% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lpocv.R
\name{evalGeneSet}
\alias{evalGeneSet}
\title{Evaluate a single gene set}
\usage{
evalGeneSet(gsi, XY, lP, nBK = 0, method = "lgr")
}
\arguments{
\item{gsi}{gene set of interest, provided as a character vector}

\item{XY}{dataset, as loaded by prepareTask()}

\item{lP}{list of pairs for cross-validation, as generated by preparePairs()}

\item{nBK}{number of background sets that should be generated alongside the gene set of interest}

\item{method}{one of {"lgr", "svm, xgb", "nn"} for logistics regression,
support vector machines, xgboost and neural nets, respectively}
}
\value{
A data frame of results, where the first row corresponds to the gene set of interest,
   and all subsequent rows correspond to the generated background sets
}
\description{
Evaluates a gene set of interest in the context of a given dataset
}
