% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lpocv.R
\name{read_gmt}
\alias{read_gmt}
\title{Load gene sets from a .gmt file}
\usage{
read_gmt(fn, iName = 1)
}
\arguments{
\item{iName}{index of the token containing gene set names
(This is typically 1 for the Broad's MSigDB sets, and 2 for PathwayCommons sets)}
}
\value{
A named list of character vectors, one per gene set
}
\description{
Parses a .gmt file and puts it into the list format
}
