% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msbb.R
\name{wrangleMSBB}
\alias{wrangleMSBB}
\title{Wrangling MSBB data}
\usage{
wrangleMSBB(destDir = "~/data/amp-ad/msbb")
}
\arguments{
\item{destDir}{Destination directory where the file is to be written}
}
\value{
Character vector of filenames of the wrangled datasets (one per brain region)
}
\description{
Wrangles MSBB molecular and clinical variables into a common data frame
}
