% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rosmap.R
\name{wrangleROSMAP}
\alias{wrangleROSMAP}
\title{Wrangling ROSMAP data}
\usage{
wrangleROSMAP(destDir = "~/data/amp-ad/rosmap")
}
\arguments{
\item{destDir}{Destination directory where the file is to be written}
}
\value{
Filename of the wrangled dataset
}
\description{
Wrangles ROSMAP molecular and clinical variables into a common data frame
}
