% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lpocv.R
\name{prepareTask}
\alias{prepareTask}
\title{Generate a binary prediction task}
\usage{
prepareTask(fn, task)
}
\arguments{
\item{fn}{Filename of a dataset wrangled by wrangleROSMAP() or wrangleMSBB()}

\item{task}{One of {"AB", "AC", "BC"}, specifying the binary task of interest}
}
\description{
Loads a previously-wrangled dataset and composes a binary prediction task
}
